+++ store-backend/src/main/java/ru/voroby/storebackend/repository/OrderDAO.java
@@ -6,7 +6,7 @@ import ru.voroby.storebackend.model.Order;
 
 import java.util.stream.Stream;
 
-public interface OrderDAO extends CrudRepository<Order, Long> {
+public interface OrderDAO extends CrudRepository<Order, Integer> {
 
   @Query("select o from Order o left join fetch o.cart where o.name = :name")
   Stream<Order> findAllByName(String name);
